Fix filename and header defs
