+++ runtime/onert/loader/trix/CMakeLists.txt
@@ -5,7 +5,7 @@ endif ()
 nnfw_find_package(TRIXEngine QUIET 2.5.0)
 if(TRIXEngine_FOUND)
   message(STATUS "ONERT frontend: Found TRIXEngine")
-  list(APPEND SOURCES src/trix_loader.cc)
+  list(APPEND SOURCES trix_loader.cc)
 else()
   message(STATUS "ONERT frontend: Failed to find TRIXEngine")
   return()
@@ -14,7 +14,6 @@ endif(TRIXEngine_FOUND)
 # Loader library name should be lib<type>_loader.so
 add_library(tvn_loader SHARED ${SOURCES})
 set_target_properties(tvn_loader PROPERTIES POSITION_INDEPENDENT_CODE ON)
-target_include_directories(tvn_loader PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
 target_link_libraries(tvn_loader PRIVATE onert_core)
 target_link_libraries(tvn_loader PRIVATE nnfw_common nnfw_coverage)
 target_link_libraries(tvn_loader PRIVATE trix_engine)
